Modeling game in order to implement a bot
It's important to create a simple state that represents how the game is going, and it has to be simple because it will be cloned so many times while the future bot will be running.
Also, to simplify the PieceBehaviour, which is, in a fact, a GameObject, will create a PieceModel, and also a PieceAction which indentifies a rotation and a position on X axis for an specific piece.