MonteCarlo Tree Search
The based of the MCTS algorithm is the tree and, also, the nodes that forms it. Each node of the tree represents a possible state of the game, starting with the root which has a state with an empty board, and also each node has a piece that is the piece that is used to arrive to that state from the previous state, which is father's state.
So, the tree will can only have nodes with pieces that had played in the real game, are being played or is sured that are gonna be played (and a regular player would know it too).